Make fresh return i32

+++ lily-core/src/tc/fresh.rs
@@ -1,5 +1,5 @@
 pub struct Fresh {
-    index: usize,
+    index: i32,
 }
 
 impl Default for Fresh {
@@ -9,11 +9,11 @@ impl Default for Fresh {
 }
 
 impl Fresh {
-    pub fn from(index: usize) -> Self {
+    pub fn from(index: i32) -> Self {
         Fresh { index }
     }
 
-    pub fn fresh(&mut self) -> usize {
+    pub fn fresh(&mut self) -> i32 {
         let index = self.index;
         self.index += 1;
         index
